FieldTracer.DoBreak — Method

DoBreak(iloc, jloc, iSize, jSize)

Check to see if we should break out of an integration.

FieldTracer.Euler! — Method

Euler!(maxstep, ds, xstart, ystart, xGrid, yGrid, ux, uy, x, y)

Simple 2D tracing using Euler's method. Super fast but not super accurate.

Arguments

maxstep::Int: max steps.

ds::Float64: step size.

xstart::Float64, ystart::Float64: starting location.

xGrid::Vector{Float64},yGrid::Vector{Float64}: actual coord system.

ux::Array{Float64,2},uy::Array{Float64,2}: field to trace through.

x::Vector{Float64},y::Vector{Float64}: x, y of result stream.

FieldTracer.Euler! — Method

Euler!(iSize, jSize, kSize, maxstep, ds, xstart, ystart, zstart,
  xGrid, yGrid, zGrid, ux, uy, uz, x, y, z)

Simple 3D tracing using Euler's method.

Arguments

maxstep::Int: max steps.

ds::Float64: step size.

xstart::Float64, ystart::Float64, zstart::Float64: starting location.

xGrid::Array{Float64,2},yGrid::Array{Float64,2},zGrid::Array{Float64,2}: actual coord system.

ux::Array{Float64,2},uy::Array{Float64,2},uz::Array{Float64,2}: field to trace through.

x::Vector{Float64},y::Vector{Float64},z::Vector{Float64}: x, y, z of result stream.

FieldTracer.RK4! — Method

RK4!(maxstep, ds, xstart, ystart, xGrid, yGrid, ux, uy, x, y)

Fast and reasonably accurate 2D tracing with 4th order Runge-Kutta method and constant step size ds.

FieldTracer.RK4! — Method

RK4!(maxstep, ds, xstart,ystart,zstart,
  xGrid,yGrid,zGrid, ux,uy,uz, x,y,z)

Fast and reasonably accurate 3D tracing with 4th order Runge-Kutta method and constant step size ds.

FieldTracer.bilin_reg — Method

bilin_reg(x, y, Q00, Q01, Q10, Q11)

Bilinear interpolation for x1,y1=(0,0) and x2,y2=(1,1) Q's are surrounding points such that Q00 = F[0,0], Q10 = F[1,0], etc.

FieldTracer.getCellID — Method

getCellID(mesh::UnstructuredMesh, x, y)

Return cell ID in the unstructured grid.

FieldTracer.grid_interp! — Method

grid_interp!(x, y, z, field, ix, iy, iz, xsize, ysize)

Interpolate a value at (x,y,z) in a field. ix,iy and iz are indexes for x,y and z locations (0-based). xsize and ysize are the sizes of field in X and Y.

FieldTracer.grid_interp! — Method

grid_interp!(x, y, field, ix, iy)

Interpolate a value at (x,y) in a field. ix and iy are indexes for x,y locations (0-based).

FieldTracer.normalize_field — Method

Create unit vectors of field.

FieldTracer.normalize_field — Method

Create unit vectors of field in normalized coordinates.

FieldTracer.select_seeds — Method

 select_seeds(x, y, nSeed=100)

Generate nSeed seeding points randomly in the grid range. If you specify nSeed, use the keyword input, otherwise it will be overloaded by the 3D version seed generation.

FieldTracer.trace2d — Method

 trace2d(fieldx, fieldy, xstart, ystart, gridx, gridy; kwargs...)

2D stream tracing in structured mesh with field in 2d array and grid in range.

FieldTracer.trace2d — Method

 trace2d(mesh::UGrid, vx, vy, xstart, ystart; maxIter=1000, maxLen=1000.)

2D stream tracing in unstructured quadrilateral and triangular mesh.

FieldTracer.trace2d_eul — Method

 trace2d_eul(fieldx, fieldy, xstart, ystart, gridx, gridy;
	 maxstep=20000, ds=0.01, gridType="ndgrid", direction="both")

Given a 2D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Euler's method. While this is faster than rk4, it is less accurate. Only valid for regular grid with coordinates' range gridx and gridy. The field can be in both meshgrid or ndgrid (default) format. Supporting direction for {"both","forward","backward"}.

FieldTracer.trace2d_eul — Method

 trace2d_eul(fieldx, fieldy, xstart, ystart, grid::CartesianGrid;
	 maxstep=20000, ds=0.01, gridType="ndgrid", direction="both")

FieldTracer.trace2d_rk4 — Method

 trace2d_rk4(fieldx, fieldy, xstart, ystart, gridx, gridy;
	 maxstep=20000, ds=0.01, gridType="ndgrid", direction="both")

Given a 2D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Runge Kutta 4. Slower than Euler, but more accurate. The higher accuracy allows for larger step sizes ds. Only valid for regular grid with coordinates' range gridx and gridy. The field can be in both meshgrid or ndgrid (default) format. Supporting direction for {"both","forward","backward"}.

FieldTracer.trace3d_eul — Method

 trace3d_eul(fieldx, fieldy, fieldz, xstart, ystart, zstart, gridx, gridy,
   gridz; maxstep=20000, ds=0.01)

Given a 3D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Euler's method. Only valid for regular grid with coordinates gridx, gridy, gridz. The field can be in both meshgrid or ndgrid (default) format.

FieldTracer.trilin_reg — Method

trilin_reg(x, y, z, Q)

Trilinear interpolation for x1,y1,z1=(0,0,0) and x2,y2,z2=(1,1,1) Q's are surrounding points such that Q000 = F[0,0,0], Q100 = F[1,0,0], etc.